requirements.txt
Flask
Flask-MySQLdb
Flask-Login
Werkzeug
python-dotenv